refactor: move related functions to their own header files
